+++ nodalia-cards-2.1.2-alpha.3.js
@@ -1,2 +1,0 @@
-import "./nodalia-cards-2.1.2-alpha.4.js";
-if(typeof window!=="undefined"){window.__NODALIA_LOADER__={"mode":"compat","pkgVersion":"2.1.2-alpha.4","contentSha256_12":"7cb0527b2790","file":"nodalia-cards-2.1.2-alpha.3.js","targetFile":"nodalia-cards-2.1.2-alpha.4.js","fallbackFile":"nodalia-cards.js"};}