+++ nodalia-cards.manifest.js
@@ -1,23 +1,23 @@
 export default {
-  "pkgVersion": "2.1.2-alpha.4",
-  "contentSha256_12": "7cb0527b2790",
+  "pkgVersion": "2.1.2-alpha.5",
+  "contentSha256_12": "ad0b740f4473",
   "file": "nodalia-cards.bundle.js",
   "loaderFile": "nodalia-cards.js",
   "hacsFile": "nodalia-cards.js",
   "compatLoaderFiles": [
     "nodalia-cards-2.1.1.js",
-    "nodalia-cards-2.1.2-alpha.3.js"
+    "nodalia-cards-2.1.2-alpha.4.js"
   ],
-  "splitCoreFile": "nodalia-cards-core-2.1.2-alpha.4.js",
-  "splitCoreSha256_12": "fb9704f55dc6",
-  "splitSuiteFile": "nodalia-cards-suite-2.1.2-alpha.4.js",
-  "splitSuiteSha256_12": "0a88b58d0fd6",
-  "editorFile": "nodalia-cards-editor-2.1.2-alpha.4.js",
-  "editorSha256_12": "7f50f338e82a"
+  "splitCoreFile": "nodalia-cards-core-2.1.2-alpha.5.js",
+  "splitCoreSha256_12": "5a5e9a01a80b",
+  "splitSuiteFile": "nodalia-cards-suite-2.1.2-alpha.5.js",
+  "splitSuiteSha256_12": "b11715e67788",
+  "editorFile": "nodalia-cards-editor-2.1.2-alpha.5.js",
+  "editorSha256_12": "6374ed8f9730"
 };
-export const pkgVersion = "2.1.2-alpha.4";
-export const contentSha256_12 = "7cb0527b2790";
+export const pkgVersion = "2.1.2-alpha.5";
+export const contentSha256_12 = "ad0b740f4473";
 export const file = "nodalia-cards.bundle.js";
-export const splitCoreFile = "nodalia-cards-core-2.1.2-alpha.4.js";
-export const splitSuiteFile = "nodalia-cards-suite-2.1.2-alpha.4.js";
-export const editorFile = "nodalia-cards-editor-2.1.2-alpha.4.js";
+export const splitCoreFile = "nodalia-cards-core-2.1.2-alpha.5.js";
+export const splitSuiteFile = "nodalia-cards-suite-2.1.2-alpha.5.js";
+export const editorFile = "nodalia-cards-editor-2.1.2-alpha.5.js";